+++ tests/tests/tests_commands.py
@@ -1,11 +1,6 @@
-from django_boost.test import TestCase
 from django.core.management import call_command
 
-
-class TestAdminSiteLog(TestCase):
-
-    def test_call_command(self):
-        call_command('adminsitelog')
+from django_boost.test import TestCase
 
 
 class TestDeleteMigrations(TestCase):